Fix gcc related errors
But gcc could not compile OIIO on release mode (some unaligned memcpy is being detected on exif write (which should be ok?)

+++ Source/Tests/Core/T_MPMCQueue.cpp
@@ -37,7 +37,7 @@ TEST(MPMCQueueTest, Stress)
             producerThreads.emplace_back([&]()
             {
                 for(uint32_t i = 0; i < itemPerThreadCount; i++)
-                    queue.Enqueue(uint32_t(i + 1));
+                    queue.Enqueue(i + 1);
             });
         }
